+++ wallet/shell.nix
@@ -1,0 +1,6 @@
+let
+  self = import ../. {};
+in (self.nix-tools.shellFor {
+  name = "cardano-wallet";
+  packages = ps: [ ps.cardano-wallet ];
+})