+++ wallet/src/Cardano/Wallet/API/Internal/Handlers.hs
@@ -8,7 +8,7 @@ import           Pos.Chain.Update (SoftwareVersion)
 
 import qualified Cardano.Wallet.API.Internal as Internal
 import           Cardano.Wallet.API.Response (APIResponse, single)
-import           Cardano.Wallet.API.V1.Types (V1, Wallet, WalletImport)
+import           Cardano.Wallet.API.V1.Types (V1, Wallet, WalletImport, BackupPhrase, MnemonicBalance)
 import           Cardano.Wallet.WalletLayer (PassiveWalletLayer)
 import qualified Cardano.Wallet.WalletLayer as WalletLayer
 
@@ -18,6 +18,7 @@ handlers w = nextUpdate       w
         :<|> postponeUpdate   w
         :<|> resetWalletState w
         :<|> importWallet     w
+        :<|> calculateMnemonic w
 
 nextUpdate :: PassiveWalletLayer IO -> Handler (APIResponse (V1 SoftwareVersion))
 nextUpdate w = do
@@ -42,3 +43,8 @@ importWallet w walletImport = do
     case res of
          Left e               -> throwM e
          Right importedWallet -> pure $ single importedWallet
+
+calculateMnemonic :: PassiveWalletLayer IO -> Maybe Bool -> BackupPhrase -> Handler (APIResponse MnemonicBalance)
+calculateMnemonic w mbool phrase = do
+  res <- liftIO $ WalletLayer.calculateMnemonic w mbool phrase
+  pure $ single res