+++ wallet/src/Cardano/Wallet/Client.hs
@@ -155,6 +155,7 @@ data WalletClient m
         :: m (Either ClientError ())
     , importWallet
         :: WalletImport -> Resp m Wallet
+    , calculateMnemonic :: Maybe Bool -> BackupPhrase -> m (Either ClientError (APIResponse MnemonicBalance))
     } deriving Generic
 
 data WalletDocClient m = WalletDocClient
@@ -286,6 +287,8 @@ natMapClient phi f wc = WalletClient
         f $ phi $ resetWalletState wc
     , importWallet =
         f . phi . importWallet wc
+    , calculateMnemonic =
+        \x -> f . phi . calculateMnemonic wc x
     }
 
 -- | Run the given natural transformation over the 'WalletClient'.